+++ src/rom_19B64.c
@@ -19,7 +19,7 @@ void sub_23954(void);
 void sub_239A4(void);
 void sub_23E18(void);
 void sub_2C9A4(void);
-void sub_4E9F0(s8);
+void sub_4E9F0(s16);
 
 void RubyBoardProcess_3A_19A20(void)
 {
